inquiry/uplink/requirements.txt
cachetools==2.0.0
click==6.7
coloredlogs==6.1
dill==0.2.6
enum34==1.1.6
future==0.16.0
futures==3.1.1
gapic-google-cloud-logging-v2==0.91.3
google-api-python-client==1.6.2
google-auth==1.0.0
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-logging==1.0.0
google-gax==0.15.9
googleapis-common-protos==1.5.2
grpcio==1.3.0
httplib2==0.10.3
humanfriendly==3.1
monotonic==1.3
oauth2client==3.0.0
ply==3.8
proto-google-cloud-logging-v2==0.91.3
protobuf==3.2.0
pyasn1==0.2.3
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
